requirements.txt
fastapi
uvicorn[standard]
openai
google-play-scraper
pydantic
python-dotenv
aiohttp
